backend/requirements.txt
anyio==4.10.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.40.0
colorama==0.4.6
cryptography==45.0.7
fastapi==0.110.0
filelock==3.19.1
firebase-admin==6.6.0
fsspec==2025.9.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httplib2==0.30.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.2
pandas==2.3.2
pillow==10.4.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.15
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.14.0
threadpoolctl==3.6.0
torch @ file:///C:/Users/anish/Downloads/wheels/torch-2.8.0%2Bcpu-cp311-cp311-win_amd64.whl#sha256=7631ef49fbd38d382909525b83696dc12a55d68492ade4ace3883c62b9fc140f
torchvision @ file:///C:/Users/anish/Downloads/wheels/torchvision-0.23.0%2Bcpu-cp311-cp311-win_amd64.whl#sha256=51603eb071d0681abc4db98b10ff394ace31f425852e8de249b91c09c60eb19a
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==15.0.1
